+++ examples/rdfExplorer.tsx
@@ -39,8 +39,7 @@ function RdfExample() {
 
     return (
         <Reactodia.Workspace ref={onMount}
-            defaultLayout={defaultLayout}
-            onIriClick={({iri}) => window.open(iri)}>
+            defaultLayout={defaultLayout}>
             <Reactodia.DefaultWorkspace
                 menu={
                     <>