+++ examples/sparql.tsx
@@ -59,8 +59,7 @@ function SparqlExample() {
 
     return (
         <Reactodia.Workspace ref={onMount}
-            defaultLayout={defaultLayout}
-            onIriClick={({iri}) => window.open(iri)}>
+            defaultLayout={defaultLayout}>
             <Reactodia.DefaultWorkspace
                 menu={<ExampleToolbarMenu />}
                 canvasWidgets={[