+++ src/workspace.ts
@@ -89,7 +89,7 @@ export {
 export { RenderingState, RenderingStateEvents, RenderingLayer } from './diagram/renderingState';
 export {
     type SharedCanvasState, SharedCanvasStateEvents, CellHighlighter, ElementDecoratorResolver,
-    FindCanvasEvent, IriClickEvent, IriClickIntent, RenameLinkToLinkStateProvider,
+    FindCanvasEvent, RenameLinkToLinkStateProvider,
 } from './diagram/sharedCanvasState';
 export { Spinner, SpinnerProps, HtmlSpinner } from './diagram/spinner';